requirements.txt
webrtcvad
openai
python-dotenv
numpy
"livekit-agents[openai]~=1.0"